Fixed error invoking man on El Capitan
Invoking the Emacs `man` command on El Capitan was giving the following error:

    error in process sentinel; can't find the manpage

The `MANPATH` environment variable is empty by default on El Capitan,
and copying this variable with exec-path-from-shell was the source
of the error.  If this variable is not copied, man pages are correctly found.

Instead of calling `(exec-path-from-shell-initialize)`, which copies
`PATH` and `MANPATH`, only `PATH` should be copied by calling
`(exec-path-from-shell-copy-env "PATH")`.